Endpoint Login & Gerando Token

+++ routes/api.php
@@ -20,37 +20,37 @@ use App\Http\Controllers\Api\V1\Auth\LoginJWTController;
 */
 
 Route::group(['prefix'=>'auth'],function(){
-    Route::post('/login',[LoginJWTController::class,'login'])->name('login');
+    Route::post('/login',[LoginJWTController::class,'login'])->name('login')->name('auth.login');
 });
 
 Route::group(['prefix'=>'real-states'],function(){
-   Route::get('/',[RealStateController::class,'index'])->name('real-states');
-   Route::post('/',[RealStateController::class,'store'])->name('real-states.store');
-   Route::get('/{id}',[RealStateController::class,'show'])->name('real-states.show');;
-   Route::put('/{id}',[RealStateController::class,'update'])->name('real-states.update');;
-   Route::delete('/{id}',[RealStateController::class,'destroy'])->name('real-states.destroy');;
+   Route::get('/',[RealStateController::class,'index'])->name('real-states.index');
+   Route::post('/',[RealStateController::class,'store'])->name('real-states.store')->middleware('jwt.auth');
+   Route::get('/{id}',[RealStateController::class,'show'])->name('real-states.show');
+   Route::put('/{id}',[RealStateController::class,'update'])->name('real-states.update')->middleware('jwt.auth');
+   Route::delete('/{id}',[RealStateController::class,'destroy'])->name('real-states.destroy')->middleware('jwt.auth');
 });
 
-Route::group(['prefix'=>'users'],function(){
-   Route::get('/',[UserController::class,'index']);
-   Route::post('/',[UserController::class,'store']);
-   Route::get('/{id}',[UserController::class,'show']);
-   Route::put('/{id}',[UserController::class,'update']);
-   Route::delete('/{id}',[UserController::class,'destroy']);
+Route::group(['prefix'=>'users','middleware'=>'jwt.auth'],function(){
+   Route::get('/',[UserController::class,'index'])->name('users.index');
+   Route::post('/',[UserController::class,'store'])->name('users.store');
+   Route::get('/{id}',[UserController::class,'show'])->name('users.show');
+   Route::put('/{id}',[UserController::class,'update'])->name('users.update');
+   Route::delete('/{id}',[UserController::class,'destroy'])->name('users.destroy');
 });
 
 Route::group(['prefix'=>'categories'],function(){
-   Route::get('/',[CategoryController::class,'index']);
-   Route::post('/',[CategoryController::class,'store']);
-   Route::get('/{id}/real-states',[CategoryController::class,'realstates']);
-   Route::get('/{id}',[CategoryController::class,'show']);
-   Route::put('/{id}',[CategoryController::class,'update']);
-   Route::delete('/{id}',[CategoryController::class,'destroy']);
+   Route::get('/',[CategoryController::class,'index'])->name('categories.index');
+   Route::post('/',[CategoryController::class,'store'])->name('categories.store')->middleware('jwt.auth');
+   Route::get('/{id}/real-states',[CategoryController::class,'realstates'])->name('categories.show.realstates');
+   Route::get('/{id}',[CategoryController::class,'show'])->name('categories.show');
+   Route::put('/{id}',[CategoryController::class,'update'])->name('categories.update')->middleware('jwt.auth');
+   Route::delete('/{id}',[CategoryController::class,'destroy'])->name('categories.delete')->middleware('jwt.auth');
 });
 
-Route::group(['prefix'=>'photos'],function(){
-   Route::put('/set-tumb/{photoId}/{realstateId}',[RealStatePhotoController::class,'setTumb']);
-   Route::delete('/remove/{id}',[RealStatePhotoController::class,'remove']);
+Route::group(['prefix'=>'photos','middleware'=>'jwt.auth'],function(){
+   Route::put('/set-tumb/{photoId}/{realstateId}',[RealStatePhotoController::class,'setTumb'])->name('photos.set-tumb');
+   Route::delete('/remove/{id}',[RealStatePhotoController::class,'remove'])->name('photos.remove');
 });
